*Core: (int64) cast bug in BlockStream
git-svn-id: svn://ultimatepp.org/upp/trunk@4185 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/BlockStream.cpp
@@ -163,7 +163,7 @@ void BlockStream::_Put(const void *data, dword size) {
 		s += n;
 		n = dword(pg1 - pg0) - pagesize;
 		streamsize = max(pos + pagesize + n, streamsize);
-		int wpos = pos + pagesize;
+		int64 wpos = pos + pagesize;
 		SetPos(pos0 + size);
 		SyncPage();
 		if(n)
